[FIX] sale_management/saas~13.3.1.0: wrong json int
Json number also includes floating point, type cast to decimal instead
of int.

upg-7276

closes odoo/upgrade#2206

+++ migrations/sale_management/saas~13.3.1.0/pre-migrate.py
@@ -16,7 +16,7 @@ def migrate(cr, version):
                    tmpl.company_id as tmpl_cid
               FROM ir_default def
               JOIN ir_model_fields field ON field.id=def.field_id
-              JOIN sale_order_template tmpl on tmpl.id=def.json_value::int
+              JOIN sale_order_template tmpl on tmpl.id=def.json_value::decimal
              WHERE field.model = 'sale.order'
                AND field.name = 'sale_order_template_id'
                AND def.user_id IS NULL
@@ -39,7 +39,7 @@ def migrate(cr, version):
             SELECT def.id
               FROM ir_default def
               JOIN ir_model_fields field ON field.id=def.field_id
-              JOIN sale_order_template tmpl on tmpl.id=def.json_value::int
+              JOIN sale_order_template tmpl on tmpl.id=def.json_value::decimal
              WHERE field.model = 'sale.order'
                AND field.name = 'sale_order_template_id'
                AND def.user_id IS NULL
